# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/vendors": {
      "post": {
        "operationId": "VendorsController_create",
        "summary": "Create a new vendor (admin only)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVendorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vendor successfully created."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "tags": [
          "vendors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "VendorsController_findAll",
        "summary": "Get all vendors with optional filtering",
        "parameters": [
          {
            "name": "country",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "service",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "Columns to select (comma-separated)",
            "schema": {
              "example": "id,name,country"
            }
          },
          {
            "name": "searchBy",
            "required": false,
            "in": "query",
            "description": "Columns to search in (comma-separated)",
            "schema": {
              "example": "name,country"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term",
            "schema": {
              "example": "vendor"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort by column and direction (e.g., name:ASC)",
            "schema": {
              "example": "id:ASC"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": 10
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vendors retrieved successfully."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "tags": [
          "vendors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/vendors/{id}": {
      "get": {
        "operationId": "VendorsController_findOne",
        "summary": "Get a vendor by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vendor retrieved successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Vendor not found."
          }
        },
        "tags": [
          "vendors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "VendorsController_update",
        "summary": "Update a vendor by ID (admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVendorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vendor updated successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Vendor not found."
          }
        },
        "tags": [
          "vendors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "VendorsController_remove",
        "summary": "Delete a vendor by ID (admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vendor deleted successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Vendor not found."
          }
        },
        "tags": [
          "vendors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/research": {
      "post": {
        "operationId": "ResearchController_create",
        "summary": "Create a research document",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResearchDocDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Research document successfully created."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "tags": [
          "research"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ResearchController_findAll",
        "summary": "Get all research documents",
        "parameters": [
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "Columns to select (comma-separated)",
            "schema": {
              "example": "title,content,createdAt"
            }
          },
          {
            "name": "searchBy",
            "required": false,
            "in": "query",
            "description": "Columns to search in (comma-separated)",
            "schema": {
              "example": "title,content"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term",
            "schema": {
              "example": "document"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort by column and direction (e.g., name:ASC)",
            "schema": {
              "example": "createdAt:DESC"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": 10
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Research documents retrieved successfully."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "tags": [
          "research"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/research/search": {
      "get": {
        "operationId": "ResearchController_search",
        "summary": "Search research documents",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Research documents retrieved successfully."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "tags": [
          "research"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/research/{id}": {
      "get": {
        "operationId": "ResearchController_findOne",
        "summary": "Get a research document by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Research document retrieved successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Research document not found."
          }
        },
        "tags": [
          "research"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ResearchController_update",
        "summary": "Update a research document",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResearchDocDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Research document updated successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Research document not found."
          }
        },
        "tags": [
          "research"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ResearchController_remove",
        "summary": "Delete a research document",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Research document deleted successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Research document not found."
          }
        },
        "tags": [
          "research"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Expansion Management API",
    "description": "API for managing client projects, vendor matching, and analytics",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateVendorDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateVendorDto": {
        "type": "object",
        "properties": {}
      },
      "CreateResearchDocDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}